Added new helper
Added a new helper to get the closest settlement. Mostly as I didn't want to write the same thing over and over.

+++ RandomEvents/RandomEvents/Events/BetMoney.cs
@@ -27,7 +27,7 @@ namespace CryingBuffalo.RandomEvents.Events
 				new InquiryElement("b", "Decline", null)
 			};
 
-			int goldToBet = (int)MathF.Floor(Hero.MainHero.Gold * moneyBetPercent);
+			int goldToBet = MathF.Floor(Hero.MainHero.Gold * moneyBetPercent);
 
 			string extraDialogue = "";
 			if (goldToBet > 40000)
